refactor: вынес общий код в один пакет
