+++ hw-5/src/main/java/ru/dankoy/hw5/core/dao/book/manytomany/BookGenreRelation.java
@@ -1,4 +1,4 @@
-package ru.dankoy.hw5.core.dao.book.mergemanytomanybycode.manytomany;
+package ru.dankoy.hw5.core.dao.book.manytomany;
 
 
 import lombok.Getter;
