--- hw-5/src/main/java/ru/dankoy/hw5/core/dao/book/mergemanytomanybycode/manytomany/BookAuthorRelation.java
@@ -1,16 +1,0 @@
-package ru.dankoy.hw5.core.dao.book.mergemanytomanybycode.manytomany;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.ToString;
-
-@Getter
-@ToString
-@RequiredArgsConstructor
-public class BookAuthorRelation {
-
-  private final long bookId;
-  private final long authorId;
-
-}
