+++ hw-5/src/main/java/ru/dankoy/hw5/core/dao/book/mergemanytomanybyjoin/BookDaoJdbcJoin.java
@@ -19,9 +19,9 @@ import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import ru.dankoy.hw5.core.dao.author.AuthorDao;
 import ru.dankoy.hw5.core.dao.book.BookDao;
-import ru.dankoy.hw5.core.dao.book.mergemanytomanybyjoin.manytomany.BookAuthorRelation;
-import ru.dankoy.hw5.core.dao.book.mergemanytomanybyjoin.manytomany.BookGenreRelation;
-import ru.dankoy.hw5.core.dao.book.mergemanytomanybyjoin.manytomany.BookResultSetExtractor;
+import ru.dankoy.hw5.core.dao.book.manytomany.BookAuthorRelation;
+import ru.dankoy.hw5.core.dao.book.manytomany.BookGenreRelation;
+import ru.dankoy.hw5.core.dao.book.mergemanytomanybyjoin.extractor.BookResultSetExtractor;
 import ru.dankoy.hw5.core.dao.genre.GenreDao;
 import ru.dankoy.hw5.core.domain.Author;
 import ru.dankoy.hw5.core.domain.Book;
