+++ hw-5/src/main/java/ru/dankoy/hw5/core/dao/book/mergemanytomanybyjoin/extractor/BookResultSetExtractor.java
@@ -1,4 +1,4 @@
-package ru.dankoy.hw5.core.dao.book.mergemanytomanybyjoin.manytomany;
+package ru.dankoy.hw5.core.dao.book.mergemanytomanybyjoin.extractor;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
